chatbot/requirements.txt
appdirs==1.4.3
APScheduler==3.6.3
asgiref==3.2.10
backcall==0.2.0
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.2
decorator==4.4.2
distlib==0.3.0
distro==1.4.0
Django==3.1.2
django-better-admin-arrayfield==1.4.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
lockfile==0.12.2
msgpack==0.6.2
packaging==20.3
parso==0.7.1
pep517==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.0
progress==1.5
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.7.2
pyparsing==2.4.6
python-telegram-bot==13.0
pytoml==0.1.21
pytz==2020.1
requests==2.22.0
retrying==1.3.3
six==1.14.0
sqlparse==0.4.1
tornado==6.0.4
traitlets==4.3.3
tzlocal==2.1
ujson==4.0.1
urllib3==1.25.8
wcwidth==0.2.5
webencodings==0.5.1
